bot/requirements.txt
python-telegram-bot==13.4.1
matplotlib==3.3.3
pymongo==3.11.3
scipy==1.6.3
typing_extensions==3.7.4.3
requests==2.25.1
pycairo==1.20.0
plotnine==0.8.0
requests_cache==0.6.3
py-staticmaps==0.4.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
timezonefinder==5.2.0
numba==0.53.1